requirements.txt
Flask==3.0.0
Jinja2==3.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
requests==2.31.0
APScheduler==3.10.4
